fix: Clippy警告39件を全て解消
## 修正内容
- clone on Copy型: `.clone()` を値コピーに変更 (15件)
- io_other_error: `io::Error::new(Other, ...)` → `io::Error::other()` (5件)
- useless_vec: `vec![]` → 配列リテラル `[]` (4件)
- unnecessary_cast: 不要なキャストを削除 (1件)
- manual_range_contains: `(0.0..=1.0).contains()` に変更 (1件)
- redundant_closure: 冗長なクロージャをメソッド参照に変更 (1件)
- iter_count: `.iter().count()` → `.len()` (1件)
- manual_is_multiple_of: `% 90 == 0` → `.is_multiple_of(90)` (1件)
- deprecated cargo_bin: `env!("CARGO_BIN_EXE_")` に変更 (1件)

## 検証
- `cargo clippy -- -D warnings`: 警告0件
- `cargo test`: 1169テスト成功 (15 ignored)
- `cargo fmt`: フォーマット適用済み

+++ superbook-pdf/src/util.rs
@@ -622,7 +622,7 @@ mod tests {
 
         let results: Vec<_> = (0..1000)
             .into_par_iter()
-            .map(|i| clamp(i as i32, 100, 900))
+            .map(|i| clamp(i, 100, 900))
             .collect();
 
         assert_eq!(results.len(), 1000);
